# api/openapi.yaml
openapi: 3.1.0
info:
  title: 慈濟 60 週年活動網站 API
  description: |
    慈濟基金會 60 週年活動網站的後端 API。
    提供主題、活動、影響力、祝福語、圖片庫等資源的 CRUD 操作。
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: 開發環境
  - url: https://api.example.com
    description: 正式環境
...
paths:
...
  /api/events:
    get:
      tags:
        - Events
      summary: 取得活動列表
      description: |
        取得活動列表，支援多種篩選方式：
        - 依月份和年份篩選: ?month=8&year=2026
        - 依主題篩選: ?topic_id=1
        - 僅依月份篩選: ?month=8
      parameters:
        - name: month
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 12
          description: 月份 (1-12)
        - name: year
          in: query
          schema:
            type: integer
          description: 年份
        - name: topic_id
          in: query
          schema:
            type: integer
          description: 主題 ID
      responses:
        '200':
          description: 成功取得活動列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
  /api/events/{id}:
    get:
      tags:
        - Events
      summary: 取得單一活動詳情
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 活動 ID
      responses:
        '200':
          description: 成功取得活動詳情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: 活動不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: integer
          description: 活動 ID
        title:
          type: string
          description: 活動標題
        description:
          type: string
          nullable: true
          description: 活動描述
        date_start:
          type: string
          format: date
          description: 活動開始日期
        date_end:
          type: string
          format: date
          nullable: true
          description: 活動結束日期
        participation_type:
          type: string
          nullable: true
          description: 參與方式
        image_url:
          type: string
          nullable: true
          description: 活動圖片 URL
        topic_id:
          type: integer
          nullable: true
          description: 所屬主題 ID
        month:
          type: integer
          minimum: 1
          maximum: 12
          description: 月份
        year:
          type: integer
          description: 年份
        sort_order:
          type: integer
          description: 排序順序
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: 錯誤訊息